fix: change address id from param to body
